requirements.txt
# Google Drive API
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1

# OpenAI for embeddings
openai==1.3.7

# Pinecone vector database
pinecone

# Document processing
PyPDF2==3.0.1
python-docx==1.1.0
python-magic==0.4.27

# Utilities
python-dotenv==1.0.0
tiktoken==0.5.1
requests==2.31.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
